requirements.txt
python-dotenv
langgraph
langchain-core
pydantic
chromadb
sentence-transformers
langchain-groq
ragas
datasets
pandas
